refactor: 에러 메서지 protocal로 처리하도록 변경, heritageListVC 축약 제거
